requirements.txt
requests
pyyaml
validators
fastapi
uvicorn
colorama
